pyproject.toml
[tool.poetry]
name = "simple-stock-management"
version = "0.1.0"
description = "Building an API for stock management"
authors = ["AlekysCoelho <alekyscoelho@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
factory-boy = "^3.3.1"
faker = "^33.1.0"
pytest-factoryboy = "^2.7.0"

[tool.pytest.ini_options]
testpaths = ["tests"] # Simplifique para a pasta raiz de testes
python_files = ["test_*.py"]
python_functions = ["test_*"]
DJANGO_SETTINGS_MODULE = "app.settings"
addopts = [
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--maxfail=5",
    "--disable-warnings",
]

[tool.coverage.run] # Seção 'run' para configurar a coleta de cobertura
branch = true # Habilita cobertura de branches (opcional, mas recomendado)
source = ["app"] # Aponte diretamente para o diretório 'app'

omit = [
    "app/migrations/*",  # Use 'app/migrations/*' para ser mais específico
    "app/tests/*",  # Exclua os testes da cobertura (importante!)
    "app/*/__init__.py", # Exclua arquivos __init__.py
    "app/asgi.py",   # Exclua arquivos específicos que não precisam de cobertura
    "app/wsgi.py",
    "manage.py" # Exclui o arquivo manage.py
]

[tool.coverage.report]
show_missing = true
skip_covered = true
threshold = 80

[tool.coverage.html]
directory = "coverage_html_report"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
